Added new module mainmenu.py with the class MainMenu. Removed the class from  main.py. Updated ref docs.
